+++ collector-bitmovin-player/src/androidTest/java/com/bitmovin/analytics/bitmovin/player/KotlinMetadataTest.kt
@@ -13,6 +13,6 @@ class KotlinMetadataTest {
     @Test
     fun verifyMetadata() {
         val metadata = KotlinClassMetadata.readLenient(BitmovinSdkAdapter::class.java.getAnnotation(Metadata::class.java)!!)
-        assertThat(metadata.version.toString()).isEqualTo("2.1.0")
+        assertThat(metadata.version.toString()).isEqualTo("2.2.0")
     }
 }