+++ collector-media3-exoplayer/src/androidTest/java/com/bitmovin/analytics/media3/exoplayer/Media3ExoPlayerConstants.kt
@@ -3,5 +3,5 @@ package com.bitmovin.analytics.media3.exoplayer
 import com.bitmovin.analytics.test.utils.PlayerInfo
 
 object Media3ExoPlayerConstants {
-    val playerInfo = PlayerInfo("media3-exoplayer", "media3-exoplayer-1.9.0", "Android:Media3")
+    val playerInfo = PlayerInfo("media3-exoplayer", "media3-exoplayer-1.10.0", "Android:Media3")
 }